Search for a value in an array of integers using binary search

+++ 0x1E-search_algorithms/1-binary.c
@@ -22,7 +22,7 @@ int binary_search(int *array, size_t size, int value)
 	while (left <= right)
 	{
 		printf("Searching in array: ");
-		for (i = left; i < right; i++)
+		for (i = left; i < (int)right; i++)
 			printf("%d, ", array[i]);
 		printf("%d\n", array[i]);
